+++ ilyvion.Laboratory/ParentNameHandlers/ICustomParentNameHandler.cs
@@ -1,0 +1,8 @@
+using System.Xml;
+
+namespace ilyvion.Laboratory.ParentNameHandlers;
+
+public interface ICustomParentNameHandler
+{
+    XmlNode? GetBestParentFor(XmlNode node, string parentNameData, List<XmlNode> allRegisteredNodes);
+}